third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-position-x-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-position-x value 'x-start'
  assert_true: 'x-start' is a supported value for background-position-x. expected true got false
[FAIL] Property background-position-x value 'x-end'
  assert_true: 'x-end' is a supported value for background-position-x. expected true got false
[FAIL] Property background-position-x value 'left -20%'
  assert_true: 'left -20%' is a supported value for background-position-x. expected true got false
[FAIL] Property background-position-x value 'right -10px'
  assert_true: 'right -10px' is a supported value for background-position-x. expected true got false
[FAIL] Property background-position-x value '0.5em, x-start, x-end'
  assert_true: '0.5em, x-start, x-end' is a supported value for background-position-x. expected true got false
[FAIL] Property background-position-x value 'calc(10px - 0.5em), left -20%, right 10px'
  assert_true: 'calc(10px - 0.5em), left -20%, right 10px' is a supported value for background-position-x. expected true got false
Harness: the test ran to completion.
